Move some stuff into the ambiguity and suggestion modules
